+++ Anarchy.Tests/Resources/FileNames.cs
@@ -1,0 +1,8 @@
+﻿namespace Anarchy.Resources;
+
+internal static class Filenames
+{
+    public const string Image1 = @".\Resources\image1.png";
+    public const string Image2 = @".\Resources\image2.jpg";
+    public const string PoetryTxt = @".\Resources\poetry.txt";
+}