+++ Anarchy/WebSockets/Gateway/Guild/Threads/ThreadMembersEventArgs.cs
@@ -14,7 +14,7 @@ namespace Discord.Gateway
         public ulong Id { get; private set; }
 
         [JsonProperty("guild_id")]
-        private ulong _guildId;
+        private readonly ulong _guildId;
         public MinimalGuild Guild => new MinimalGuild(_guildId).SetClient(Client);
 
         [JsonProperty("member_count")]